+++ BE/src/main/java/team05/codesquad/issuetracker/controller/milestonedto/MilestoneDto.java
@@ -1,4 +1,4 @@
-package team05.codesquad.issuetracker.dto;
+package team05.codesquad.issuetracker.controller.milestonedto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
