+++ BE/src/main/java/team05/codesquad/issuetracker/controller/milestonedto/request/MilestoneCreateRequest.java
@@ -1,4 +1,4 @@
-package team05.codesquad.issuetracker.dto.request;
+package team05.codesquad.issuetracker.controller.milestonedto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,10 +7,11 @@ import lombok.extern.slf4j.Slf4j;
 
 import java.time.LocalDate;
 
-@Slf4j @Getter
+@Slf4j
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class MilestoneUpdateRequest {
+public class MilestoneCreateRequest {
 
     private String title;
     private String description;
