+++ BE/src/main/java/team05/codesquad/issuetracker/controller/milestonedto/response/MilestoneCreateResponse.java
@@ -1,4 +1,4 @@
-package team05.codesquad.issuetracker.dto.response;
+package team05.codesquad.issuetracker.controller.milestonedto.response;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
