--- BE/src/main/java/team05/codesquad/issuetracker/dto/response/MilestoneUpdateResponse.java
@@ -1,18 +1,0 @@
-package team05.codesquad.issuetracker.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.extern.slf4j.Slf4j;
-
-import java.time.LocalDate;
-
-@Slf4j
-@Getter
-@Builder
-public class MilestoneUpdateResponse {
-
-    private String title;
-    private String description;
-    private LocalDate deadLine;
-
-}
